requirements.txt
# Backend Dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
faker>=20.0.0
reportlab>=4.0.0
pdfplumber>=0.10.0
faiss-cpu>=1.7.0
openai>=1.0.0
openai-agents>=0.2.9
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
weasyprint>=60.0.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
pypdf>=3.17.0
tenacity>=8.2.0

# Testing Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Frontend Dependencies
gradio>=4.0.0
